requirements.txt
sentence-transformers
pinecone
python-dotenv
uvicorn
fastapi

